Update Report.ipynb
Riboflavin, also known as vitamin B2, was found by England chemist Bruce in 1879. However, the chemical essence of this substance was initially not detected until 1933, when American chemist Berger extracted and named this chemical compound as riboflavin. Riboflavin is the universal precursor of flavin mononucleotide (FMN) and flavin adenine dinucleotide (FAD), both of which act as hydrogen carriers and are essential for the activity of a wide variety of metabolic enzymes in higher eukaryotes Riboflavin has an important effect on respiration because it is a coenzyme of flavin protein. This substance participates directly in the biological oxidation of protein, fat, and carbohydrate and make an important influence on life activity. At the same time, riboflavin is the essential nutrient of metabolism and physiological functions. Riboflavin is synthesized by all plants, fungi and most bacteria, but not by higher animals including humans. Therefore, humans and animals must obtain riboflavin through dietary sources. People need 0.3-1.8g riboflavin per day to maintain a healthy body condition. Also, riboflavin accumulates cell regeneration, the growth of skin, nails, and hair. Riboflavin is an important food additive, medicine, and food dye. Additionally, riboflavin is found to cure migraine, angina pectoris, and cancer. The product of riboflavin will increase by 4.5% during the period from 2020 to 2025. There are four international companies (e.g., BASF and DSM) that focus on producing riboflavin available in the form of vitamins, botanicals, and fatty acids.